+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Removed SDX common clone in docker
   - Restore defaults for qids 130, 131, 132.
   - Remove pip `--require-hashes` requirement
+  - Make use of sequence list endpoint
   - Populate default dates on certain surveys from metadata
   - Add fix for cryptography install in docker